+++ scripts/deployers/StakingRewards.ts
@@ -5,9 +5,11 @@ const { ethers } = hre;
 
 export const deployStakingRewards = async (owner: string, rUmb: string = CONFIG.stage1.rUmb.address!) => {
   const Contract = await ethers.getContractFactory('StakingRewards');
-  const contract = await Contract.deploy(owner, CONFIG.stage1.farming.rewardsDistribution, CONFIG.UMB.address, rUmb);
+  const rewardsDistribution = CONFIG.stage1.farming.rewardsDistribution || CONFIG.multiSig.address
+  const contract = await Contract.deploy(owner, rewardsDistribution, CONFIG.UMB.address, rUmb);
 
   await contract.deployed();
   console.log('deployed StakingRewards:', contract.address);
+  console.log('reward distributor is:', rewardsDistribution);
   return contract;
 };